refactoring mpnn to mpnn module
